chore: remove accessoverviewrework flag
