+++ frontend/src/component/admin/users/AccessOverview/AccessOverview.tsx
@@ -3,23 +3,20 @@ import { PageHeader } from 'component/common/PageHeader/PageHeader';
 import { useRequiredPathParam } from 'hooks/useRequiredPathParam';
 import useUserInfo from 'hooks/api/getters/useUserInfo/useUserInfo';
 import { styled, useMediaQuery } from '@mui/material';
-import { useEffect, useMemo, useState } from 'react';
+import { useMemo, useState } from 'react';
 import { useEnvironments } from 'hooks/api/getters/useEnvironments/useEnvironments';
 import { ConditionallyRender } from 'component/common/ConditionallyRender/ConditionallyRender';
 import theme from 'themes/theme';
-import { StringParam, useQueryParams } from 'use-query-params';
 import useProjects from 'hooks/api/getters/useProjects/useProjects';
-import { AccessOverviewSelect } from './AccessOverviewSelect.tsx';
 import { useUserAccessOverview } from 'hooks/api/getters/useUserAccessOverview/useUserAccessOverview';
-import { AccessOverviewAccordion } from './AccessOverviewAccordion/AccessOverviewAccordion.tsx';
-import {
-    getCategorizedProjectPermissions,
-    getCategorizedRootPermissions,
-} from 'utils/permissions';
+import { getCategorizedRootPermissions } from 'utils/permissions';
 import type { IAccessOverviewPermissionCategory } from './AccessOverviewAccordion/AccessOverviewList.tsx';
-import { createProjectPermissionsStructure } from 'component/admin/roles/RoleForm/RolePermissionCategories/createProjectPermissionsStructure';
 import { Search } from 'component/common/Search/Search';
 import { SearchHighlightProvider } from 'component/common/Table/SearchHighlightContext/SearchHighlightContext';
+import { RootAccessOverviewAccordion } from './AccessOverviewAccordion/RootAccessOverviewAccordion.tsx';
+import { RootRoleGroupAccess } from './RootRoleGroupAccess.tsx';
+import { filterCategory } from './ProjectAccess/ProjectAccess.tsx';
+import { ProjectAccessSection } from './ProjectAccess/ProjectAccessSection.tsx';
 
 const StyledActionsContainer = styled('div')(({ theme }) => ({
     display: 'flex',
@@ -41,26 +38,12 @@ const StyledAccessOverviewContainer = styled('div')(({ theme }) => ({
     gap: theme.spacing(2),
 }));
 
-const filterCategory = (
-    category: IAccessOverviewPermissionCategory,
-    search: string,
-): IAccessOverviewPermissionCategory | undefined => {
-    const searchLower = search.toLowerCase();
-    const filteredPermissions = category.permissions.filter(({ displayName }) =>
-        displayName.toLowerCase().includes(searchLower),
-    );
-
-    if (filteredPermissions.length) {
-        return { ...category, permissions: filteredPermissions };
-    }
-};
+const StyledProjectAccessWrapper = styled('div')(({ theme }) => ({
+    marginTop: theme.spacing(4),
+}));
 
 export const AccessOverview = () => {
     const id = useRequiredPathParam('id');
-    const [query, setQuery] = useQueryParams({
-        project: StringParam,
-        environment: StringParam,
-    });
     const { projects } = useProjects();
     const { environments } = useEnvironments();
     const { user, loading } = useUserInfo(id);
@@ -68,48 +51,14 @@ export const AccessOverview = () => {
     const isSmallScreen = useMediaQuery(theme.breakpoints.down('md'));
 
     const [searchValue, setSearchValue] = useState('');
-    const [project, setProject] = useState(query.project ?? '');
-    const [environment, setEnvironment] = useState(
-        query.environment ?? undefined,
-    );
 
-    const { overview, rootRole, projectRoles } = useUserAccessOverview(
-        id,
-        project,
-        environment,
-    );
+    const { overview, rootRole } = useUserAccessOverview(id);
 
-    useEffect(() => {
-        setQuery(
-            {
-                project: project || undefined,
-                environment,
-            },
-            'replace',
-        );
-    }, [project, environment]);
+    const rootRoleGroups = overview?.groups.filter((r) => r.rootRole);
 
     const AccessActions = (
         <StyledActionsContainer>
             <Search initialValue={searchValue} onChange={setSearchValue} />
-            <AccessOverviewSelect
-                label='Project'
-                options={projects}
-                getOptionLabel={(option) => option?.name ?? ''}
-                value={projects.find(({ id }) => id === project)}
-                setValue={(value) => setProject(value?.id ?? '')}
-            />
-            <AccessOverviewSelect
-                label='Environment'
-                options={environments}
-                getOptionLabel={(option) =>
-                    option?.name.concat(
-                        !option.enabled ? ' - deprecated' : '',
-                    ) ?? ''
-                }
-                value={environments.find(({ name }) => name === environment)}
-                setValue={(value) => setEnvironment(value?.name)}
-            />
         </StyledActionsContainer>
     );
 
@@ -124,29 +73,6 @@ export const AccessOverview = () => {
             .map((category) => filterCategory(category, searchValue))
             .filter(Boolean) as IAccessOverviewPermissionCategory[];
     }, [overview?.root, searchValue]);
-
-    const projectCategories = useMemo(() => {
-        const categories = createProjectPermissionsStructure(
-            overview?.project ?? [],
-        ).map(({ label, permissions }) => ({
-            label,
-            permissions: permissions.map(([permission]) => permission),
-        })) as IAccessOverviewPermissionCategory[];
-
-        return categories
-            .map((category) => filterCategory(category, searchValue))
-            .filter(Boolean) as IAccessOverviewPermissionCategory[];
-    }, [overview?.project, searchValue]);
-
-    const environmentCategories = useMemo(() => {
-        const categories = getCategorizedProjectPermissions(
-            overview?.environment ?? [],
-        ) as IAccessOverviewPermissionCategory[];
-
-        return categories
-            .map((category) => filterCategory(category, searchValue))
-            .filter(Boolean) as IAccessOverviewPermissionCategory[];
-    }, [overview?.environment, searchValue]);
 
     return (
         <PageContent
@@ -170,22 +96,20 @@ export const AccessOverview = () => {
         >
             <StyledAccessOverviewContainer>
                 <SearchHighlightProvider value={searchValue}>
-                    <AccessOverviewAccordion categories={rootCategories}>
-                        Root permissions for role {rootRole?.name}
-                    </AccessOverviewAccordion>
-                    <AccessOverviewAccordion categories={projectCategories}>
-                        Project permissions
-                        {project
-                            ? ` for project ${project}${projectRoles?.length ? ` with project role${projectRoles.length !== 1 ? 's' : ''} ${projectRoles?.map((role: any) => role.name).join(', ')}` : ''}`
-                            : ''}
-                    </AccessOverviewAccordion>
-                    {environment && (
-                        <AccessOverviewAccordion
-                            categories={environmentCategories}
-                        >
-                            Environment permissions for {environment}
-                        </AccessOverviewAccordion>
-                    )}
+                    <RootAccessOverviewAccordion
+                        rootRole={rootRole}
+                        groups={rootRoleGroups}
+                        categories={rootCategories}
+                    />
+                    <RootRoleGroupAccess groups={rootRoleGroups ?? []} />
+                    <StyledProjectAccessWrapper>
+                        <ProjectAccessSection
+                            id={id}
+                            projects={projects}
+                            environments={environments}
+                            searchValue={searchValue}
+                        />
+                    </StyledProjectAccessWrapper>
                 </SearchHighlightProvider>
             </StyledAccessOverviewContainer>
         </PageContent>
